requirements.txt
requests == 2.26.0
Django==2.1
azure_storage==0.36.0
djangorestframework==3.11.0
azure-cosmos==4.2.0
whitenoise==5.0.1
django-oauth-toolkit==1.3.2
mysqlclient==1.4.6
pandas==1.0.3
pymysql==0.9.3